requirements.txt
streamlit
numpy
pandas
matplotlib
scikit-learn
scipy
python-dotenv